buy pokemon

+++ backend/PokemonTraderApi/Src/User/Models.cs
@@ -2,9 +2,9 @@ namespace PokemonTraderApi.User;
 
 public class PokemonUser
 {
-  public int pokemonUserId { get; set; }
+  public long pokemonUserId { get; set; }
   public string? username { get; set; }
-  public int coins { get; set; }
+  public long coins { get; set; }
 
   public PokemonUser() { }
   public PokemonUser(Microsoft.AspNetCore.Identity.IdentityUser user)
